+++ tests/unit/utilities/src/math_parser_service.cpp
@@ -9,9 +9,9 @@ ZTEST_SUITE(math_parser_service, NULL, NULL, NULL, NULL, NULL);
 ZTEST(math_parser_service, test_MathParserService_Evaluate_returns_correct_value) {
     MathParserService math_parser_service;
     math_parser_service.SetExpression("(x + y) * 4");
-    double x = 8.0;
+    float x = 8.0;
     math_parser_service.DefineVariable("x", &x);
-    double y = 2.0;
+    float y = 2.0;
     math_parser_service.DefineVariable("y", &y);
 
     zassert_equal(math_parser_service.Evaluate(), 40.0);
@@ -20,9 +20,9 @@ ZTEST(math_parser_service, test_MathParserService_Evaluate_returns_correct_value
 ZTEST(math_parser_service, test_MathParserService_evaluator_reuse) {
     MathParserService math_parser_service;
     math_parser_service.SetExpression("(x + y) * 4");
-    double x = 8.0;
+    float x = 8.0;
     math_parser_service.DefineVariable("x", &x);
-    double y = 2.0;
+    float y = 2.0;
     math_parser_service.DefineVariable("y", &y);
 
     zassert_equal(math_parser_service.Evaluate(), 40.0);
@@ -30,9 +30,9 @@ ZTEST(math_parser_service, test_MathParserService_evaluator_reuse) {
     math_parser_service.SetExpression("(x - 8 * var_d) / f");
     x = 80.0;
     math_parser_service.DefineVariable("x", &x);
-    double var_d = 2.0;
+    float var_d = 2.0;
     math_parser_service.DefineVariable("var_d", &var_d);
-    double f = 2.0;
+    float f = 2.0;
     math_parser_service.DefineVariable("f", &f);
 
     zassert_equal(math_parser_service.Evaluate(), 32.0);
@@ -43,9 +43,9 @@ ZTEST(math_parser_service, test_MathParserService_missing_vars_fail) {
     MathParserService math_parser_service;
     math_parser_service.SetExpression("(x + y) * 4");
 
-    double x = 8.0;
+    float x = 8.0;
     math_parser_service.DefineVariable("x", &x);
-    double y = 2.0;
+    float y = 2.0;
     math_parser_service.DefineVariable("y", &y);
 
     zassert_equal(math_parser_service.Evaluate(), 40.0);
